+++ src/core/idp/event_handlers.js
@@ -234,15 +234,13 @@ export async function handleMessageFromQueue(messageStr) {
   }
 }
 
-export async function handleTendermintNewBlockEvent(
+export async function handleTendermintNewBlock(
   error,
-  result,
+  height,
   missingBlockCount
 ) {
   if (missingBlockCount == null) return;
   try {
-    const height = tendermint.getBlockHeightFromNewBlockEvent(result);
-
     // messages that arrived before 'NewBlock' event
     // including messages between the start of missing block's height
     // and the block before latest block height
@@ -338,7 +336,7 @@ export async function handleTendermintNewBlockEvent(
     logger.error(err.getInfoForLog());
     await common.notifyError({
       callbackUrl: callbackUrls.error_url,
-      action: 'handleTendermintNewBlockEvent',
+      action: 'handleTendermintNewBlock',
       error: err,
     });
   }