+++ src/utils/external_crypto_service.js
@@ -339,6 +339,7 @@ async function callbackWithRetry(url, body, logPrefix) {
           details: {
             url,
             cbId,
+            callbackRetryTimedOut: true,
           },
         });
       }